move handle into common
